+++ src/actions/AppActions.js
@@ -62,7 +62,9 @@ export function _loadDataFromFile(path, options) {
         }));
 
         try {
-            await dispatch(loadSettingsFromFile(join(getUserDataPath(), 'coreSettings.json'), true));
+            if (!options.skipSettings) {
+                await dispatch(loadSettingsFromFile(join(getUserDataPath(), 'coreSettings.json'), true));
+            }
 
             if (!path) {
                 path = getState().settings.dataFolder;
@@ -98,7 +100,8 @@ export function _loadDataFromFile(path, options) {
         } catch (error) {
             dispatch(updateProcess({
                 id: processId,
-                state: 'ERROR'
+                state: 'ERROR',
+                error: error.toString()
             }));
 
             throw error;
@@ -158,7 +161,8 @@ export function _loadDataFromServer(options) {
         } catch (error) {
             dispatch(updateProcess({
                 id: processId,
-                state: 'ERROR'
+                state: 'ERROR',
+                error: error.toString()
             }));
 
             throw error;
@@ -244,7 +248,8 @@ export function _saveDataToFile(path, options) {
         } catch (error) {
             dispatch(updateProcess({
                 id: processId,
-                state: 'ERROR'
+                state: 'ERROR',
+                error: error.toString()
             }));
 
             throw error;
@@ -286,7 +291,8 @@ export function cleanData() {
         } catch (error) {
             dispatch(updateProcess({
                 id: processId,
-                state: 'ERROR'
+                state: 'ERROR',
+                error: error.toString()
             }));
 
             throw error;
@@ -381,7 +387,8 @@ export function resetData(options) {
         } catch (error) {
             dispatch(updateProcess({
                 id: processId,
-                state: 'ERROR'
+                state: 'ERROR',
+                error: error.toString()
             }));
 
             throw error;