+++ src/actions/BackupActions.js
@@ -97,7 +97,8 @@ export function cleanBackups() {
         } catch (error) {
             dispatch(updateProcess({
                 id: processId,
-                state: 'ERROR'
+                state: 'ERROR',
+                error: error.toString()
             }));
 
             throw error;